LPS-144724 Abstract the OpenSocialBookmarkPropsTransformer
Avoiding repeat the same implementation in each module
