+++ modules/apps/social/social-bookmarks-taglib/src/main/resources/META-INF/resources/index.js
@@ -12,4 +12,4 @@
  * details.
  */
 
-export {default as openSocialBookmark} from './js/openSocialBookmark';
+export {default as OpenSocialBookmarkPropsTransformer} from './js/OpenSocialBookmarkPropsTransformer';
