+++ modules/apps/social/social-bookmarks-taglib/src/main/resources/META-INF/resources/js/OpenSocialBookmarkPropsTransformer.js
@@ -12,4 +12,19 @@
  * details.
  */
 
-export {OpenSocialBookmarkPropsTransformer as default} from 'social-bookmarks-taglib';
+import openSocialBookmark from './openSocialBookmark';
+
+export default function propsTransformer({
+	additionalProps: {className, classPK, postURL, type, url},
+	...otherProps
+}) {
+	return {
+		...otherProps,
+		onClick(event) {
+			event.preventDefault();
+			event.stopPropagation();
+
+			return openSocialBookmark({className, classPK, postURL, type, url});
+		},
+	};
+}
